make maxStaleAfter uint16

+++ packages/core/contracts/types/ProtocolParameter.sol
@@ -33,7 +33,7 @@ struct ProtocolParameter {
     uint256 maxStaleAfter;
 }
 struct StoredProtocolParameter {
-    /* slot 0 (29) */
+    /* slot 0 (28) */
     uint24 maxFee;                  // <= 1677%
     uint32 maxLiquidationFee;       // <= 4294
     uint24 maxCut;                  // <= 1677%
@@ -42,7 +42,7 @@ struct StoredProtocolParameter {
     uint24 minEfficiency;           // <= 1677%
     uint24 referralFee;             // <= 1677%
     uint24 minScale;                // <= 1677%
-    uint24 maxStaleAfter;           // <= 4660 hours
+    uint16 maxStaleAfter;           // <= 18 hours
 }
 struct ProtocolParameterStorage { StoredProtocolParameter value; } // SECURITY: must remain at (1) slots
 using ProtocolParameterStorageLib for ProtocolParameterStorage global;
@@ -63,7 +63,7 @@ library ProtocolParameterStorageLib {
             UFixed6.wrap(uint256(value.minEfficiency)),
             UFixed6.wrap(uint256(value.referralFee)),
             UFixed6.wrap(uint256(value.minScale)),
-            uint24(value.maxStaleAfter)
+            uint16(value.maxStaleAfter)
         );
     }
 
@@ -81,7 +81,7 @@ library ProtocolParameterStorageLib {
         if (newValue.maxRate.gt(UFixed6.wrap(type(uint32).max / 2))) revert ProtocolParameterStorageInvalidError();
         if (newValue.minMaintenance.gt(UFixed6.wrap(type(uint24).max))) revert ProtocolParameterStorageInvalidError();
         if (newValue.minEfficiency.gt(UFixed6.wrap(type(uint24).max))) revert ProtocolParameterStorageInvalidError();
-        if (newValue.maxStaleAfter > uint256(type(uint24).max)) revert ProtocolParameterStorageInvalidError();
+        if (newValue.maxStaleAfter > uint256(type(uint16).max)) revert ProtocolParameterStorageInvalidError();
 
         self.value = StoredProtocolParameter(
             uint24(UFixed6.unwrap(newValue.maxFee)),
@@ -92,7 +92,7 @@ library ProtocolParameterStorageLib {
             uint24(UFixed6.unwrap(newValue.minEfficiency)),
             uint24(UFixed6.unwrap(newValue.referralFee)),
             uint24(UFixed6.unwrap(newValue.minScale)),
-            uint24(newValue.maxStaleAfter)
+            uint16(newValue.maxStaleAfter)
         );
     }
 }
